input_wav/requirements.txt
boto3
ulid-py
numpy
matplotlib